+++ theme/index.ts
@@ -1,0 +1,8 @@
+export { ThemeProvider } from './ThemeProvider';
+export { usePalette } from './palette/usePalette';
+export { useScale } from './scale/useScale';
+export { useSpectrum } from './spectrum/useSpectrum';
+
+export * from './palette/types';
+export * from './scale/types';
+export * from './spectrum/types';